make it all work

+++ staticpages/urls.py
@@ -9,7 +9,7 @@ rel_linksdir = os.path.join('staticpages', 'links')
 abs_links_dir = os.path.join(os.path.dirname(__file__), 'templates', rel_linksdir)
 
 def get_urls(sections=None):
-    STATICPAGES_TEMPLATES = settings.getattr('STATICPAGES_TEMPLATES')
+    STATICPAGES_TEMPLATES = getattr(settings, 'STATICPAGES_TEMPLATES')
     if sections is not None and isinstance(sections, basestring):
         sections = [sections]
     if sections is None:
@@ -17,7 +17,8 @@ def get_urls(sections=None):
     else:
         sections = [section for section in STATICPAGES_TEMPLATES if section[0] in sections]
     urlpatterns = []
-    shutil.rmtree(abs_links_dir)
+    if os.path.exists(abs_links_dir):
+        shutil.rmtree(abs_links_dir)
     os.makedirs(abs_links_dir)
     for name, path in sections:
         os.symlink(path, os.path.join(abs_links_dir, name))
@@ -29,7 +30,11 @@ def get_urls(sections=None):
                 if f.endswith('.html'):
                     filepath = os.path.join(reldir.lstrip('/'), f)
                     urlpath = os.path.splitext(filepath)[0]
-                    url_re = r'^%s/$' % urlpath
+                    if f == 'index.html':
+                        urlpath = urlpath[:-len('index')]
+                        url_re = r'^%s$' % urlpath
+                    else:
+                        url_re = r'^%s/$' % urlpath
                     template_name = os.path.join(rel_linksdir, name, filepath)
                     view = TemplateView.as_view(template_name=template_name)
                     urlpattern = url_re, view
